# [Perf] [CPU] eliminate redundant memory access in group query attention