@MaggotHATE 2025-05-05
Just tested this out of curiosity: Qwen 3 degrades in quality (ignores `/no_think`, for example), Mistral Small 2 outputs empty characters. Does it break compatibility with older models? Windows 10, i7-8700, CPU backend only.