@Excigma 2025-05-12
Hm, I opened your PR in my editor and saw this:
![image](https://github.com/user-attachments/assets/6fbc151d-d60d-48b6-affb-14569f068f19)

I ran `editorconfig-checker`, removed the whitespace, and ran it again, and it seems like the error is gone. The line that was flagged was `7058` (running checker locally) instead of `7045` (CI):
![image](https://github.com/user-attachments/assets/66d83b25-fe69-482c-ae1f-d8315b91e821)

<details>
<summary>Here's a patch to fix the change, if you can't find the line</summary>

```patch
From 3c7b2ed48acfcb5a9c06846ed0b548b3e48707af Mon Sep 17 00:00:00 2001
From: Excigma <git@excigma.xyz>
Date: Mon, 12 May 2025 15:46:03 +1200
Subject: [PATCH] style: remove trailing whitespace

---
 ggml/src/ggml-cpu/ops.cpp | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

diff --git a/ggml/src/ggml-cpu/ops.cpp b/ggml/src/ggml-cpu/ops.cpp
index 250b6abc..a1481d9e 100644
--- a/ggml/src/ggml-cpu/ops.cpp
+++ b/ggml/src/ggml-cpu/ops.cpp
@@ -7055,7 +7055,7 @@ static void ggml_compute_forward_flash_attn_ext_f16(
 
             const float * pq = (const float *) ((char *) q->data + (iq1*nbq1 + (iq2 + i_gqa)*nbq2 + iq3*nbq3));
             q_to_vec_dot(pq, Q_q[i_gqa], DK);
-  
+
             const uint32_t h = iq2 + i_gqa;
             slope[i_gqa] = (max_bias > 0.0f) ? h < n_head_log2 ? powf(m0, h + 1) : powf(m1, 2*(h - n_head_log2) + 1) : 1.0f;
         }
-- 
2.49.0

```

</details>