@ZelinMa557 2025-05-16
> Hm, I opened your PR in my editor and saw this: ![image](https://private-user-images.githubusercontent.com/30280397/442576933-6fbc151d-d60d-48b6-affb-14569f068f19.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDczNzYzMDEsIm5iZiI6MTc0NzM3NjAwMSwicGF0aCI6Ii8zMDI4MDM5Ny80NDI1NzY5MzMtNmZiYzE1MWQtZDYwZC00OGI2LWFmZmItMTQ1NjlmMDY4ZjE5LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNTA1MTYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjUwNTE2VDA2MTMyMVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTA1OWEzMmFjZmU0ZDg5OThjOTgyMzg2YzllZGVkMzAxODk0MjI3YjZhMDllOTg5MTQwNjRmNGZjNmYzMWRmOGImWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.wiqNSuaZAW-dE89yXxaZbyk184fBr1XqybFg1xLldoQ)
> 
> I ran `editorconfig-checker`, removed the whitespace, and ran it again, and it seems like the error is gone. The line that was flagged was `7058` (running checker locally) instead of `7045` (CI): ![image](https://private-user-images.githubusercontent.com/30280397/442577236-66d83b25-fe69-482c-ae1f-d8315b91e821.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDczNzYzMDEsIm5iZiI6MTc0NzM3NjAwMSwicGF0aCI6Ii8zMDI4MDM5Ny80NDI1NzcyMzYtNjZkODNiMjUtZmU2OS00ODJjLWFlMWYtZDgzMTViOTFlODIxLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNTA1MTYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjUwNTE2VDA2MTMyMVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWYzMzY4YzFmODMyZmJmOWQ5NTM3ZWYxZDE1OGFjOTE3NzM4ZWY2ODVmM2ZkMzk1N2UzYzNkOTdlODFjZGUyMjcmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.0Zz4mpqbI9rULwcJmM1edXhoJPcnNg8NGBcs5-qE1BA)
> 
> Here's a patch to fix the change, if you can't find the line

Thanks for your reply! It's true that line 7045 has some problems, I will fix it.