@slaren 2025-06-10
I tried this, but the performance I see on my system is not always better, and in some cases it is worse than master. The problem with testing with `llama-cli` is that it is very susceptible to random variations, and the results are not always repeatable. Instead, you can use the `llama-bench` and `test-backend-ops` tools to test the performance. It would be useful if you can provide results with these tools that show a clear improvement on your system.

You can try these commands to get started with these tools:
```
llama-bench -m <model.gguf> -fa 1 -p 128 -n 32 -d 0,128
test-backend-ops -b CPU -o FLASH_ATTN_EXT perf
```